refactor: 주문 도메인 테스트 코드실패 케이스 메꿈

+++ src/main/java/wayc/backend/order/domain/repository/query/OrderQueryRepository.java
@@ -4,7 +4,7 @@ import wayc.backend.order.domain.repository.query.dto.FindOrderResponseDto;
 import wayc.backend.order.domain.repository.query.dto.FindPagingOrderResponseDto;
 
 public interface OrderQueryRepository {
-    FindPagingOrderResponseDto findCustomerOrderLineItemsWithPaging(Long memberId, Integer lastOrderLineItemId);
-    FindPagingOrderResponseDto findSellerOrderLineItemsWithPaging(Long memberId, Integer lastOrderLineItemId);
+    FindPagingOrderResponseDto findCustomerOrderLineItemsWithPaging(Long memberId, Long lastOrderLineItemId);
+    FindPagingOrderResponseDto findSellerOrderLineItemsWithPaging(Long memberId, Long lastOrderLineItemId);
     FindOrderResponseDto findDetailOrderLineItem(Long orderId);
 }
